+++ HISTORY.rst
@@ -49,3 +49,4 @@ History
 * StereoNet now accept Vec3 or groups of Vec3
 * Ortensor updated with new properties E1,E2,E3
   and Vollmer(1989) parameters P,G,R and C
+* uniform_lin and uniform_fol improved.